% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_augment_by_activity.R
\name{augment_by_activity}
\alias{augment_by_activity}
\title{Receive the processed data and calculate additional statistics.}
\arguments{
\item{df}{Dataframe resulting from processed directory of .gpx files.}
}
\value{
Augmented dataframe contains cumulative distance, 
  mile and kilometer splits
}
\description{
Receive the processed data and calculate additional statistics.
}
