% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_splits.R
\name{calc_splits}
\alias{calc_splits}
\title{Calculate splits by type}
\usage{
calc_splits(cum_distance, split_type = "mile")
}
\arguments{
\item{cum_distance}{ordered vector of cumulative distances}

\item{split_type}{denotes distance at which split occurs}
}
\value{
cumulative distances between successive coordinates.
}
\description{
Calculate splits by type
}
