% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpx_list_to_df.R
\name{gpx_list_to_df}
\alias{gpx_list_to_df}
\title{Converts list of tracks into a dataframe}
\usage{
gpx_list_to_df(trk_lst)
}
\arguments{
\item{trk_lst}{list of tracks extracted from .gpx file(s)}
}
\value{
dataframe containing track data
}
\description{
Converts list of tracks into a dataframe
}
