% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_points.R
\name{parse_points}
\alias{parse_points}
\title{Extract data associated with <trkpt> nodes.}
\usage{
parse_points(pts)
}
\arguments{
\item{pts}{nodelist of <trkpt> nodes.}
}
\value{
a list containing the latitude, longitude, elevation
  and time associated with the <trkpt> node.
}
\description{
Extract data associated with <trkpt> nodes.
}
