% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_tracks.R
\name{parse_tracks}
\alias{parse_tracks}
\title{Find xml nodes and extract attributes}
\usage{
parse_tracks(trks, file_id)
}
\arguments{
\item{trks}{nodelist of <trk> nodes.}

\item{file_id}{index of file being parsed in directory.}
}
\value{
a list containing the track name, track id, activity type
  and (most importantly) the track points.
}
\description{
Find xml nodes and extract attributes
}
