% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cum_distance.R
\name{calc_cum_distance}
\alias{calc_cum_distance}
\title{Calculate the cumulative distance between points.}
\usage{
calc_cum_distance(lon, lat)
}
\arguments{
\item{lon}{ordered vector of longitude marks.}

\item{lat}{ordered vector of latitude marks.}
}
\value{
cumulative distances between successive coordinates.
}
\description{
Calculate the cumulative distance between points.
}
